Refactor puzzle schema handling and improve chat integration
Refactors puzzle and unit schemas to use Pydantic models with extra='forbid', ensuring strict schema validation. Updates puzzle creation and update logic to use model attributes instead of dicts, and improves LLM client handling for structured data. Enhances chat and editor integration with better session and puzzle refresh logic, adds new static assets for UI, and clarifies prompt instructions for puzzle generation. Also includes bug fixes for editor state reset and robust handling of chat context in puzzle export.

+++ app/schemas/node_schema.py
@@ -17,11 +17,11 @@ class NodeCreate(BaseModel):
 class NodeGenerate(BaseModel):
     model_config = ConfigDict(
         from_attributes=True,
-        extra='forbid'
+        extra="forbid"
     )
     index: int
-    x: float
-    y: float
+    x: int
+    y: int
 
 
 class NodeRead(NodeCreate):
